Refactor test files and clean code
